+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [PR178](https://github.com/lerna-stack/akka-entity-replication/pull/178)
 - Add diagnostic info to logs of sending replication results
   [PR#179](https://github.com/lerna-stack/akka-entity-replication/pull/179)
+- Persist EntitySnapshot as an event
+  [PR#184](https://github.com/lerna-stack/akka-entity-replication/pull/184)
 - Allow only specific actors defined as a sticky leader to become a leader
   [PR#186](https://github.com/lerna-stack/akka-entity-replication/pull/186)